Move the higher level mapping functions (the ones that aren't synchronicity wrapped)
