+++ modal/parallel_map.py
@@ -6,12 +6,23 @@ from typing import Any, Callable, Dict, Optional, Set, Tuple
 from aiostream import pipe, stream
 from grpclib import Status
 
-from modal._utils.async_utils import queue_batch_iterator, synchronize_api, synchronizer
+from modal._utils.async_utils import (
+    AsyncOrSyncIteratable,
+    queue_batch_iterator,
+    synchronize_api,
+    synchronizer,
+    warn_if_generator_is_not_consumed,
+)
 from modal._utils.blob_utils import BLOB_MAX_PARALLELISM
-from modal._utils.function_utils import current_input_id
+from modal._utils.function_utils import (
+    ATTEMPT_TIMEOUT_GRACE_PERIOD,
+    OUTPUTS_TIMEOUT,
+    _create_input,
+    _process_result,
+    current_input_id,
+)
 from modal._utils.grpc_utils import retry_transient_errors
 from modal.config import logger
-from modal.functions import ATTEMPT_TIMEOUT_GRACE_PERIOD, OUTPUTS_TIMEOUT, _create_input, _process_result
 from modal_proto import api_pb2
 
 if typing.TYPE_CHECKING:
@@ -222,3 +233,175 @@ async def _map_invocation(
         async for response in streamer:
             if response is not None:
                 yield response.value
+
+
+# note that `map()` is not synchronicity-wrapped, since it accepts executable code in the form of
+# iterators that we don't want to run inside the synchronicity thread. We delegate to `._map()` with
+# a safer Queue as input
+@warn_if_generator_is_not_consumed(function_name="Function.map")
+def _map_sync(
+    self,
+    *input_iterators: typing.Iterable[Any],  # one input iterator per argument in the mapped-over function/generator
+    kwargs={},  # any extra keyword arguments for the function
+    order_outputs: bool = True,  # return outputs in order
+    return_exceptions: bool = False,  # propagate exceptions (False) or aggregate them in the results list (True)
+) -> AsyncOrSyncIteratable:
+    """Parallel map over a set of inputs.
+
+    Takes one iterator argument per argument in the function being mapped over.
+
+    Example:
+    ```python
+    @stub.function()
+    def my_func(a):
+        return a ** 2
+
+
+    @stub.local_entrypoint()
+    def main():
+        assert list(my_func.map([1, 2, 3, 4])) == [1, 4, 9, 16]
+    ```
+
+    If applied to a `stub.function`, `map()` returns one result per input and the output order
+    is guaranteed to be the same as the input order. Set `order_outputs=False` to return results
+    in the order that they are completed instead.
+
+    `return_exceptions` can be used to treat exceptions as successful results:
+
+    ```python
+    @stub.function()
+    def my_func(a):
+        if a == 2:
+            raise Exception("ohno")
+        return a ** 2
+
+
+    @stub.local_entrypoint()
+    def main():
+        # [0, 1, UserCodeException(Exception('ohno'))]
+        print(list(my_func.map(range(3), return_exceptions=True)))
+    ```
+    """
+
+    return AsyncOrSyncIteratable(
+        _map_async(
+            self, *input_iterators, kwargs=kwargs, order_outputs=order_outputs, return_exceptions=return_exceptions
+        ),
+        nested_async_message="You can't iter(Function.map()) or Function.for_each() from an async function. Use async for ... Function.map.aio() or Function.for_each.aio() instead.",
+    )
+
+
+@warn_if_generator_is_not_consumed(function_name="Function.map.aio")
+async def _map_async(
+    self,
+    *input_iterators: typing.Union[
+        typing.Iterable[Any], typing.AsyncIterable[Any]
+    ],  # one input iterator per argument in the mapped-over function/generator
+    kwargs={},  # any extra keyword arguments for the function
+    order_outputs: bool = True,  # return outputs in order
+    return_exceptions: bool = False,  # propagate exceptions (False) or aggregate them in the results list (True)
+) -> typing.AsyncGenerator[Any, None]:
+    """mdmd:hidden
+    This runs in an event loop on the main thread
+
+    It concurrently feeds new input to the input queue and yields available outputs
+    to the caller.
+    Note that since the iterator(s) can block, it's a bit opaque how often the event
+    loop decides to get a new input vs how often it will emit a new output.
+    We could make this explicit as an improvement or even let users decide what they
+    prefer: throughput (prioritize queueing inputs) or latency (prioritize yielding results)
+    """
+    raw_input_queue: Any = SynchronizedQueue()  # type: ignore
+    raw_input_queue.init()
+
+    async def feed_queue():
+        # This runs in a main thread event loop, so it doesn't block the synchronizer loop
+        async with stream.zip(*[stream.iterate(it) for it in input_iterators]).stream() as streamer:
+            async for args in streamer:
+                await raw_input_queue.put.aio((args, kwargs))
+        await raw_input_queue.put.aio(None)  # end-of-input sentinel
+
+    feed_input_task = asyncio.create_task(feed_queue())
+
+    try:
+        async for output in self._map.aio(raw_input_queue, order_outputs, return_exceptions):  # type: ignore[reportFunctionMemberAccess]
+            yield output
+    finally:
+        feed_input_task.cancel()  # should only be needed in case of exceptions
+
+
+def _for_each_sync(self, *input_iterators, kwargs={}, ignore_exceptions: bool = False):
+    """Execute function for all inputs, ignoring outputs.
+
+    Convenient alias for `.map()` in cases where the function just needs to be called.
+    as the caller doesn't have to consume the generator to process the inputs.
+    """
+    # TODO(erikbern): it would be better if this is more like a map_spawn that immediately exits
+    # rather than iterating over the result
+    for _ in self.map(*input_iterators, kwargs=kwargs, order_outputs=False, return_exceptions=ignore_exceptions):
+        pass
+
+
+async def _for_each_async(self, *input_iterators, kwargs={}, ignore_exceptions: bool = False):
+    async for _ in self.map.aio(  # type: ignore
+        *input_iterators, kwargs=kwargs, order_outputs=False, return_exceptions=ignore_exceptions
+    ):
+        pass
+
+
+@warn_if_generator_is_not_consumed(function_name="Function.starmap")
+async def _starmap_async(
+    self,
+    input_iterator: typing.Union[typing.Iterable[typing.Sequence[Any]], typing.AsyncIterable[typing.Sequence[Any]]],
+    kwargs={},
+    order_outputs: bool = True,
+    return_exceptions: bool = False,
+) -> typing.AsyncIterable[Any]:
+    raw_input_queue: Any = SynchronizedQueue()  # type: ignore
+    raw_input_queue.init()
+
+    async def feed_queue():
+        # This runs in a main thread event loop, so it doesn't block the synchronizer loop
+        async with stream.iterate(input_iterator).stream() as streamer:
+            async for args in streamer:
+                await raw_input_queue.put.aio((args, kwargs))
+        await raw_input_queue.put.aio(None)  # end-of-input sentinel
+
+    feed_input_task = asyncio.create_task(feed_queue())
+    try:
+        async for output in self._map.aio(raw_input_queue, order_outputs, return_exceptions):  # type: ignore[reportFunctionMemberAccess]
+            yield output
+    finally:
+        feed_input_task.cancel()  # should only be needed in case of exceptions
+
+
+@warn_if_generator_is_not_consumed(function_name="Function.starmap.aio")
+def _starmap_sync(
+    self,
+    input_iterator: typing.Iterable[typing.Sequence[Any]],
+    kwargs={},
+    order_outputs: bool = True,
+    return_exceptions: bool = False,
+) -> AsyncOrSyncIteratable:
+    """Like `map`, but spreads arguments over multiple function arguments.
+
+    Assumes every input is a sequence (e.g. a tuple).
+
+    Example:
+    ```python
+    @stub.function()
+    def my_func(a, b):
+        return a + b
+
+
+    @stub.local_entrypoint()
+    def main():
+        assert list(my_func.starmap([(1, 2), (3, 4)])) == [3, 7]
+    ```
+    """
+    return AsyncOrSyncIteratable(
+        _starmap_async(
+            self, input_iterator, kwargs=kwargs, order_outputs=order_outputs, return_exceptions=return_exceptions
+        ),
+        nested_async_message="You can't run Function.map() or Function.for_each() from an async function. Use Function.map.aio()/Function.for_each.aio() instead.",
+    )
